toolkit/content/tests/chrome/bug366992_window.xul

<?xml version="1.0"?>

<!-- ***** BEGIN LICENSE BLOCK *****
   - Version: MPL 1.1/GPL 2.0/LGPL 2.1
   -
   - The contents of this file are subject to the Mozilla Public License Version
   - 1.1 (the "License"); you may not use this file except in compliance with
   - the License. You may obtain a copy of the License at
   - http://www.mozilla.org/MPL/
   -
   - Software distributed under the License is distributed on an "AS IS" basis,
   - WITHOUT WARRANTY OF ANY KIND, either express or implied. See the License
   - for the specific language governing rights and limitations under the
   - License.
   -
   - The Original Code is bug 366992 unit test
   -
   - The Initial Developer of the Original Code is
   - Mozilla Corporation.
   - Portions created by the Initial Developer are Copyright (C) 2007
   - the Initial Developer. All Rights Reserved.
   -
   - Contributor(s):
   -     Asaf Romano <mano@mozilla.com>
   -
   - Alternatively, the contents of this file may be used under the terms of
   - either the GNU General Public License Version 2 or later (the "GPL"), or
   - the GNU Lesser General Public License Version 2.1 or later (the "LGPL"),
   - in which case the provisions of the GPL or the LGPL are applicable instead
   - of those above. If you wish to allow use of your version of this file only
   - under the terms of either the GPL or the LGPL, and not to allow others to
   - use your version of this file under the terms of the MPL, indicate your
   - decision by deleting the provisions above and replace them with the notice
   - and other provisions required by the GPL or the LGPL. If you do not delete
   - the provisions above, a recipient may use your version of this file under
   - the terms of any one of the MPL, the GPL or the LGPL.
   -
   - ***** END LICENSE BLOCK ***** -->

<?xml-stylesheet href="chrome://global/skin" type="text/css"?>
<?xul-overlay href="chrome://global/content/editMenuOverlay.xul"?>

<window id="366992 test"
        xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul"
        onload="onLoad();"
        width="600"
        height="600"
        title="366992 test">

  <commandset id="editMenuCommands"/>

  <script type="application/javascript"
          src="chrome://global/content/globalOverlay.js"/>
  <script type="application/javascript"><![CDATA[
    // Without the fix for bug 366992, the copy command would be enabled
    // for the textbox even though the textbox's controller for this command
    // disables it.
    var gShouldNotBeReachedController = {
      supportsCommand: function(aCommand) {
        return aCommand == "cmd_copy";
      },
      isCommandEnabled: function(aCommand) {
        return aCommand == "cmd_copy";
      },
      doCommand: function(aCommand) { }
    }

    function ok(condition, message) {
      window.opener.wrappedJSObject.SimpleTest.ok(condition, message);
    }
    function finish() {
      window.close();
      window.opener.wrappedJSObject.SimpleTest.finish();
    }

    function onLoad() {
      document.getElementById("textbox").focus();
      var copyDisabled = document.getElementById("cmd_copy")
                                 .getAttribute("disabled") == "true";
      ok(copyDisabled,
         "cmd_copy should be disabled when the empty textbox is focused");
      finish();
    }
    
    window.controllers.appendController(gShouldNotBeReachedController);
  ]]></script>

  <textbox id="textbox"/>
</window>
